add conftest.py
