+++ tests/test_package_gen.py
@@ -3,24 +3,6 @@
 import subprocess
 from pathlib import Path
 
-import pytest
-
-
-@pytest.fixture()
-def project_config():
-    """
-    Pytest fixture for defining the project config.
-
-    Returns
-    -------
-    dict
-        dictionary with values for the cookiecutter template,
-        as defined in the cookiecutter.json
-    """
-    return {
-        "project_slug": "cookiecutter_test",
-    }
-
 
 def test_package_generation(
     tmp_path: Path,
